docker-steps/step-dataset-val-config/requirements.txt
absl-py==2.0.0
asttokens==2.4.0
backcall==0.2.0
bio==1.5.9
biopython==1.81
biothings-client==0.3.0
boto3==1.28.63
botocore==1.31.63
cachetools==5.3.1
certifi==2023.7.22
charset-normalizer==3.3.0
checksumdir==1.2.0
click==8.1.7
colorama==0.4.6
comm==0.1.4
contourpy==1.1.1
cycler==0.12.1
decorator==5.1.1
executing==2.0.0
filelock==3.12.4
fonttools==4.25.0
fsspec==2023.9.2
future==0.18.3
google-auth==2.23.3
google-auth-oauthlib==1.0.0
gprofiler-official==1.0.0
grpcio==1.59.0
huggingface-hub==0.17.3
idna==3.4
importlib-metadata==6.8.0
importlib-resources==6.1.0
ipython==8.12.3
ipywidgets==8.1.1
jedi==0.19.1
Jinja2==3.1.2
jmespath==1.0.1
joblib==1.3.2
jupyterlab-widgets==3.0.9
kiwisolver==1.4.5
Markdown==3.5
MarkupSafe==2.1.3
matplotlib==3.7.3
matplotlib-inline==0.1.6
mpmath==1.3.0
munkres==1.1.4
mygene==3.2.2
networkx==3.1
numpy==1.24.4
oauthlib==3.2.2
packaging==23.2
pandas==2.0.3
parso==0.8.3
pickleshare==0.7.5
Pillow==10.0.1
platformdirs==3.11.0
ply==3.11
pooch==1.7.0
prompt-toolkit==3.0.39
protobuf==3.20.3
pure-eval==0.2.2
pyasn1==0.5.0
pyasn1-modules==0.3.0
Pygments==2.16.1
pyparsing==3.1.1
PyQt5==5.15.7
PyQt5-Qt5==5.15.2
PyQt5-sip==12.13.0
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
PyYAML==6.0.1
regex==2023.10.3
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
s3transfer==0.7.0
sacremoses==0.0.53
safetensors==0.4.0
scikit-learn==1.3.1
scipy==1.10.1
sentencepiece==0.1.91
six==1.16.0
stack-data==0.6.3
sympy==1.12
tabulate==0.9.0
tensorboard==2.14.0
tensorboard-data-server==0.7.1
threadpoolctl==3.2.0
tokenizers==0.14.1
torch==2.1.0
tqdm==4.66.1
traitlets==5.11.2
typing_extensions==4.8.0
tzdata==2023.3
urllib3==1.26.17
wcwidth==0.2.8
Werkzeug==3.0.0
widgetsnbextension==4.0.9
zipp==3.17.0
